Added 4 Glints mode and super basic homography stuffs

+++ eyeWriterTracker/src/Tracking/eyeFinder.h
@@ -16,78 +16,78 @@
 #include "ofxVectorMath.h"
 
 class eyeFinder  {
-
+	
 public:
-
+	
 	void setup(int width, int height, int _targetWidth, int _targetHeight, int divisor = 1);
-	bool update(ofxCvGrayscaleAdvanced & _currentImg, int threshold, int minBlobSize, int maxBlobSize, bool bAllArea);
+	bool update(ofxCvGrayscaleAdvanced & _currentImg, float threshold, float minBlobSize, float maxBlobSize, bool bAllArea);
 	void drawFindingRect(float x, float y, float width, float height);
 	void getRotatedBoundingBox(int blobNum);
-
+	
 	ofxCvContourFinder			contourFinder;
 	ofxCvGrayscaleAdvanced		currentImg;
 	ofxCvGrayscaleAdvanced		previousImg;
 	ofxCvGrayscaleAdvanced		diffImg;
-
-//!!!: only to check
-//	ofxCvGrayscaleAdvanced		tempCurrent;
-//	ofxCvGrayscaleAdvanced		tempPrevious;
-//	ofxCvGrayscaleAdvanced		temptest_dif;
-
+	
+	//!!!: only to check
+	//	ofxCvGrayscaleAdvanced		tempCurrent;
+	//	ofxCvGrayscaleAdvanced		tempPrevious;
+	//	ofxCvGrayscaleAdvanced		temptest_dif;	
+	
 	ofxCvBlob						foundBlob;
-
+	
 	float			w,h,div;
 	float			wdiv, hdiv;
 	int				targetWidth;
 	int				targetHeight;
-
+	
 	ofRectangle		offsetROI;
 	ofRectangle		boundingRect;
 	ofPoint			centroid;
-
+	
 	bool			bBadFrame;
 	bool			bFoundOne;
 	bool			bFirstFrame;
-
+	
 	// Filter
-
+	
 	bool			bUseGamma;
 	float			gamma;
-
+	
 	bool			bUseDilate;
 	int				nErosions;
 	int				nDilations;
-
+	
 	bool			bUseBlur;
 	float			blur;
-
+	
 	bool			bUseContrast;
-
+	
 	float			contrast;
 	float			brightness;
-
+	
 	// Shape Filter
-
+	
 	bool			bUseCompactnessTest;
 	float			maxCompactness;
-
+	
 	bool			bUseLongShortTest;
 	float			minLongShortRatio;
-	CvBox2D		box;
-
+	CvBox2D			box;
+	
 	float			maxWidth;
 	float			maxHeight;
-
+	
 	float			minSquareness;
-
-
+	
+	
 private:
-
+	
 	ofRectangle	findingRect;
-
-
-
-
+	
+	
+	
+	
 };
 
 #endif /* _EYE_FINDER_ */
